Made corrections to the project, made utility functions to the module
