+++ src/components/filter.js
@@ -10,6 +10,7 @@ const createFilterMarkup = (filter, isChecked) => {
       class="filter__input visually-hidden"
       name="filter"
       ${isChecked ? `checked` : ``}
+      ${count ? `` : `disabled`}
     />
     <label for="filter__${title}" class="filter__label">
     ${title} <span class="filter__${title}-count">${count}</span></label
