+++ src/utils.js
@@ -10,3 +10,17 @@ export const formatTime = (date) => {
 
   return `${hours}:${minutes}`;
 };
+
+// Получает случаное целое число в заданном диапазоне
+const getRandomIntegerNumber = (min, max) => {
+  return min + Math.floor(Math.random() * (max - min));
+};
+
+// Возвращает случайный элемент массива
+const getRandomArrayItem = (array) => {
+  const randomIndex = getRandomIntegerNumber(0, array.length);
+
+  return array[randomIndex];
+};
+
+export {getRandomIntegerNumber, getRandomArrayItem};
